// LMS_Frontend/src/app/models/sendBook.model.ts
// FOR USING NORMAL APP

export interface sendBook {
    bookName: string;
    bookISBNNo: string;
    bookAuthor: string;
    bookPublisher: string;
    bookDescription: string;
    bookCategory: string;
    bookPrice: number;
    bookEdition: string;
    bookImageUrl: string;
  }

  // FOR USING iTURMERIC
  // export interface sendBook {
  //   bookId:string;
  //     bookName: string;
  //     bookISBNNo: string;
  //     bookAuthor: string;
  //     bookPublisher: string;
  //     bookDescription: string;
  //     bookCategory: string;
  //     bookPrice: number;
  //     bookEdition: string;
  //     bookImageUrl: string;
  //   }
  
  
    


  